--- ml-agents/mlagents/trainers/common/optimizer/__init__.py
@@ -1,1 +1,0 @@
-from mlagents.trainers.common.optimizer.optimizer import Optimizer  # noqa
