--- ml-agents/mlagents/trainers/common/policy/__init__.py
@@ -1,1 +1,0 @@
-from mlagents.trainers.common.policy.policy import Policy  # noqa
