--- ml-agents/mlagents/trainers/common/trainer/__init__.py
@@ -1,1 +1,0 @@
-from mlagents.trainers.common.trainer.trainer import Trainer  # noqa
