+++ ml-agents/mlagents/trainers/components/reward_signals/curiosity/model.py
@@ -2,7 +2,7 @@ from typing import List, Tuple
 from mlagents.tf_utils import tf
 
 from mlagents.trainers.models import ModelUtils
-from mlagents.trainers.common.policy.tf_policy import TFPolicy
+from mlagents.trainers.policy.tf_policy import TFPolicy
 
 
 class CuriosityModel(object):
