+++ ml-agents/mlagents/trainers/components/reward_signals/gail/model.py
@@ -2,7 +2,7 @@ from typing import List, Optional, Tuple
 
 from mlagents.tf_utils import tf
 
-from mlagents.trainers.common.policy.tf_policy import TFPolicy
+from mlagents.trainers.policy.tf_policy import TFPolicy
 from mlagents.trainers.models import ModelUtils
 
 EPSILON = 1e-7
