+++ ml-agents/mlagents/trainers/optimizer/__init__.py
@@ -1,0 +1,1 @@
+from mlagents.trainers.optimizer.optimizer import Optimizer  # noqa
