+++ ml-agents/mlagents/trainers/policy/__init__.py
@@ -1,0 +1,1 @@
+from mlagents.trainers.policy.policy import Policy  # noqa
