+++ ml-agents/mlagents/trainers/policy/nn_policy.py
@@ -8,8 +8,8 @@ from mlagents_envs.base_env import BatchedStepResult
 from mlagents.trainers.brain import BrainParameters
 from mlagents.trainers.models import EncoderType
 from mlagents.trainers.models import ModelUtils
-from mlagents.trainers.common.policy.tf_policy import TFPolicy
-from mlagents.trainers.common.distributions import (
+from mlagents.trainers.policy.tf_policy import TFPolicy
+from mlagents.trainers.distributions import (
     GaussianDistribution,
     MultiCategoricalDistribution,
 )
