+++ ml-agents/mlagents/trainers/tests/test_bcmodule.py
@@ -5,7 +5,7 @@ import numpy as np
 import yaml
 import os
 
-from mlagents.trainers.common.policy.nn_policy import NNPolicy
+from mlagents.trainers.policy.nn_policy import NNPolicy
 from mlagents.trainers.components.bc.module import BCModule
