+++ ml-agents/mlagents/trainers/trainer/__init__.py
@@ -1,0 +1,1 @@
+from mlagents.trainers.trainer.trainer import Trainer  # noqa
