+++ ml-agents/mlagents/trainers/trainer/rl_trainer.py
@@ -3,9 +3,9 @@ import logging
 from typing import Dict
 from collections import defaultdict
 
-from mlagents.trainers.common.optimizer.tf_optimizer import TFOptimizer
+from mlagents.trainers.optimizer.tf_optimizer import TFOptimizer
 from mlagents.trainers.buffer import AgentBuffer
-from mlagents.trainers.common.trainer import Trainer
+from mlagents.trainers.trainer import Trainer
 from mlagents.trainers.exception import UnityTrainerException
 from mlagents.trainers.components.reward_signals import RewardSignalResult
